pyproject.toml
[tool.poetry]
name = "clientai-dungeon-master"
version = "0.1.0"
description = "An AI-powered dungeon master for text-based RPG adventures"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "clientai_dungeon_master"}]

[tool.poetry.dependencies]
python = "^3.11"
clientai = "^0.1.2"
requests = "^2.32.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
